[nextest-runner] move unit output-related code into a file
